made a helper for the report page tests and added the redux setup to report page tests
